fixed the no history data when the event.local_polcies is empty
