+++ manager/pkg/cronjob/task/compliance_history_test.go
@@ -12,7 +12,8 @@ import (
 )
 
 var _ = Describe("sync the compliance data", Ordered, func() {
-	const sourceTable = "event.local_policies"
+	const localPolicyEventTable = "event.local_policies"
+	const localStatusComplianceTable = "local_status.compliance"
 	const targetTable = "history.local_compliance"
 
 	BeforeAll(func() {
@@ -42,7 +43,14 @@ var _ = Describe("sync the compliance data", Ordered, func() {
 			EXCEPTION
 				WHEN duplicate_object THEN null;
 			END $$;
-
+			CREATE TABLE IF NOT EXISTS local_status.compliance (
+				id uuid NOT NULL,
+				cluster_name character varying(63) NOT NULL,
+				leaf_hub_name character varying(63) NOT NULL,
+				error status.error_type NOT NULL,
+				compliance local_status.compliance_type NOT NULL,
+				cluster_id uuid
+			);
 			CREATE TABLE IF NOT EXISTS event.local_policies (
 				policy_id uuid NOT NULL,
 				cluster_id uuid NOT NULL,
@@ -80,8 +88,9 @@ var _ = Describe("sync the compliance data", Ordered, func() {
 			}
 			defer rows.Close()
 			expectedTableSet := map[string]bool{
-				sourceTable: true,
-				targetTable: true,
+				localStatusComplianceTable: true,
+				localPolicyEventTable:      true,
+				targetTable:                true,
 			}
 
 			for rows.Next() {
@@ -100,7 +109,7 @@ var _ = Describe("sync the compliance data", Ordered, func() {
 		}, 10*time.Second, 2*time.Second).ShouldNot(HaveOccurred())
 	})
 
-	It("sync the data from the source table to the target", func() {
+	It("sync the data from the event.local_policies to the history.local_compliance", func() {
 		By("Create the sync job")
 		s := gocron.NewScheduler(time.UTC)
 		complianceJob, err := s.Every(1).Second().Tag("LocalCompliance").DoWithJobDetails(
@@ -186,6 +195,63 @@ var _ = Describe("sync the compliance data", Ordered, func() {
 			}
 			if logCount < 1 {
 				return fmt.Errorf("table history.local_compliance_job_log records are not synced")
+			}
+			return nil
+		}, 10*time.Second, 2*time.Second).ShouldNot(HaveOccurred())
+	})
+
+	It("sync the data from local_status.compliance to history.local_compliance", func() {
+		By("Create the sync job")
+		s := gocron.NewScheduler(time.UTC)
+		complianceJob, err := s.Every(1).Second().Tag("LocalCompliance").DoWithJobDetails(
+			task.SyncLocalCompliance, ctx, pool, true)
+		Expect(err).ToNot(HaveOccurred())
+		fmt.Println("set local compliance job", "scheduleAt", complianceJob.ScheduledAtTime())
+		s.StartAsync()
+		defer s.Clear()
+		By("Create the data to the source table")
+		_, err = pool.Exec(ctx, `
+					INSERT INTO local_status.compliance (id, cluster_name, leaf_hub_name, error, compliance, cluster_id) VALUES
+					('f8c4479f-fec5-44d8-8060-da9a92d5e138', 'local_cluster1', 'leaf1', 'none', 'compliant',
+					'0cd723ab-4649-42e8-b8aa-aa094ccf06b4'),
+					('f8c4479f-fec5-44d8-8060-da9a92d5e139', 'local_cluster2', 'leaf2', 'none', 'compliant',
+					'0cd723ab-4649-42e8-b8aa-aa094ccf06b4'),
+					('f8c4479f-fec5-44d8-8060-da9a92d5e137', 'local_cluster3', 'leaf3', 'none', 'compliant',
+					'0cd723ab-4649-42e8-b8aa-aa094ccf06b4');
+				`)
+		Expect(err).ToNot(HaveOccurred())
+
+		By("Check whether the data is copied to the target table")
+		Eventually(func() error {
+			rows, err := pool.Query(ctx, `
+					SELECT id, cluster_id, compliance,compliance_changed_frequency
+					FROM history.local_compliance`)
+			if err != nil {
+				return err
+			}
+			defer rows.Close()
+			syncCount := 0
+			expectRecordMap := map[string]string{
+				"f8c4479f-fec5-44d8-8060-da9a92d5e138": "compliant",
+				"f8c4479f-fec5-44d8-8060-da9a92d5e139": "compliant",
+				"f8c4479f-fec5-44d8-8060-da9a92d5e137": "compliant",
+			}
+			fmt.Println("finding simulating records:")
+			for rows.Next() {
+				var id, cluster_id, compliance string
+				var compliance_changed_frequency int
+				err := rows.Scan(&id, &cluster_id, &compliance, &compliance_changed_frequency)
+				if err != nil {
+					return err
+				}
+				fmt.Println(id, cluster_id, compliance, compliance_changed_frequency)
+				if status, ok := expectRecordMap[id]; ok && status == compliance && compliance_changed_frequency == 0 {
+					syncCount++
+					delete(expectRecordMap, id)
+				}
+			}
+			if len(expectRecordMap) > 0 {
+				return fmt.Errorf("table history.local_compliance records are not synced")
 			}
 			return nil
 		}, 10*time.Second, 2*time.Second).ShouldNot(HaveOccurred())
